Fix: fixed openpyxl versions
