+++ setup.py
@@ -11,6 +11,7 @@ setup(
     package_dir={'': 'src'},
     install_requires=[
         'pandas>=1.3.5',
+        'openpyxl>=3.1.5',
         'google-cloud-speech>=2.26.0',
         'google-cloud-storage>=2.16.0',
         'gspread>=6.1.2',
